backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.4
langgraph==0.0.66
langchain-openai==0.1.17
jinja2==3.1.4
pytest==9.0.2
pytest-asyncio==0.24.0

